+++ tools/clang/test/HLSLFileCheck/dxil/debug/noop_no_fold_int.hlsl
@@ -8,22 +8,23 @@ Texture2D tex1 : register(t1);
 [RootSignature("DescriptorTable(SRV(t0), SRV(t1))")]
 float4 main() : SV_Target {
   // CHECK: %[[preserve_i32:[0-9]+]] = load i32, i32* @dx.preserve.value
-  // CHECK: %[[preserve_f32:[0-9]+]] = sitofp i32 %[[preserve_i32]]
 
-  float x = 10;
-  // fadd float 1.000000e+01, %[[preserve_f32]]
+  int x = 10;
+  // or i32 10, %[[preserve_i32]]
 
-  float y = x + 5;
-  // CHECK: %[[a1:.+]] = fadd
-  // fadd float [[a1]], %[[preserve_f32]]
+  int y = x + 5;
+  // CHECK: %[[a1:.+]] = add
+  // or i32 [[a1]], %[[preserve_i32]]
 
-  float z = y * 2;
-  // CHECK: %[[b1:.+]] = fmul
-  // fadd float [[b1]], %[[preserve_f32]]
+  int z = y * 2;
+  // CHECK: %[[b1:.+]] = mul
+  // or i32 [[b1]], %[[preserve_f32]]
 
-  float w = z / 0.5;
-  // CHECK: %[[c1:.+]] = fdiv
-  // fadd float [[c1]], %[[preserve_f32]]
+  int w = z / 0.5;
+  // CHECK: sitofp
+  // CHECK: fdiv
+  // CHECK: %[[c1:.+]] = fptosi
+  // or i32 [[c1]], %[[preserve_f32]]
 
   Texture2D tex = tex0; 
   // CHECK: load i32, i32* @dx.nothing
